refactor(Pages): Associated adapters with pages
