+++ pages/adapters/ContentBlockAdapter.tsx
@@ -2,18 +2,21 @@ import React from 'react'
 import camelCase from 'lodash/camelCase'
 import { documentToReactComponents } from '@contentful/rich-text-react-renderer'
 
-import { ComponentWithClass } from '../../../../common/ComponentWithClass'
-import { ContentPanel, LeftCol } from '../../ContentPanel'
-import { CodeBlock } from '..'
+import { ComponentWithClass } from '../../common/ComponentWithClass'
+import {
+  ContentPanel,
+  LeftCol,
+  RightCol,
+} from '../../components/presenters/ContentPanel'
 
-interface CodeBlockAdapterProps extends ComponentWithClass {
+interface ContentBlockAdapterProps extends ComponentWithClass {
   fields: any
 }
 
-export const CodeBlockAdapter: React.FC<CodeBlockAdapterProps> = ({
+export const ContentBlockAdapter: React.FC<ContentBlockAdapterProps> = ({
   fields,
 }) => {
-  const { title, description, sourceCode, filename } = fields
+  const { title, description } = fields
   const id = camelCase(title)
 
   return (
@@ -22,7 +25,7 @@ export const CodeBlockAdapter: React.FC<CodeBlockAdapterProps> = ({
         {title && <h2>{title}</h2>}
         {description?.json && documentToReactComponents(description.json)}
       </LeftCol>
-      <CodeBlock filename={filename}>{sourceCode}</CodeBlock>
+      <RightCol />
     </ContentPanel>
   )
 }
